API films funcional, controladores finalizados

+++ src/main/java/upchi/api/movie/film/controllers/dtos/responses/GetFilmResponse.java
@@ -9,6 +9,6 @@ public class GetFilmResponse {
     private Long id;
     private String title;
     private String img;
-    private Integer duration;
+    private Float duration;
 
 }
