+++ src/main/java/upchi/api/movie/film/controllers/dtos/responses/PostFilmResponse.java
@@ -8,6 +8,6 @@ public class PostFilmResponse {
     
     private Long id;
     private String title;
-    private Integer duration;
+    private Float duration;
 
 }
